Replace convention plugin classes with kts files
