--- gradle-conventions/src/main/kotlin/KonditionPublicationExtension.kt
@@ -1,3 +1,0 @@
-open class KonditionPublicationExtension {
-    var artifactId: String = ""
-}
